pyproject.toml
[project]
name = "lscolors"
version = "1.0.0"
description = "dircolors utilities"
authors = [
    {name = "Russel Lane", email = "russel@rlane.com"},
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["dircolors", "xterm-256color", "color-scheme", "terminal", "shell", "ls", "tree"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Console Fonts",
    "Topic :: System :: Filesystems",
    "Topic :: System :: Shells",
    "Topic :: Terminals",
    "Topic :: Utilities"
]
requires-python = ">=3.10"
dependencies = [
    "ansicolors>=1.1.8",
    "pyyaml>=6.0.2",
    "rlane-libcli>=1.0.4",
]

[project.urls]
Homepage = "https://github.com/russellane/lscolors"

[project.scripts]
lscolors = "lscolors.cli:main"

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.8.0",
    "flake8-bugbear>=24.8.19",
    "flake8-pydocstyle>=0.2.4",
    "flake8-pylint>=0.2.1",
    "flake8-pytest-style>=2.0.0",
    "flake8-simplify>=0.21.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mypy>=1.11.2",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3",
]

[tool.black]
line-length = "97"

[tool.isort]
line_length = "97"
profile = "black"

[tool.pylint.format]
max-line-length = "97"

[tool.pylint.messages_control]
enable = [
    "useless-suppression",
]
fail-on = [
    # useless-suppression
    "I0021",
]

disable = [
    "R0801",                    # Similar lines in 2 files - cleo commands
]

[tool.pydocstyle]
convention = "google"
add-ignore = [
    # "D105",     # pep257 numpy google   # Missing docstring in magic method
    "D105",
    # "D202",     # pep257 numpy google   # No blank lines allowed after function docstring
    "D202",
]

[tool.mypy]
strict = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# vim: set ts=4 sw=4 et:
